refactored pydantic and main app modules
